Test cases improved
