+++ test-files/expected-outputs/hacked_toy64.expected
@@ -1,0 +1,1 @@
+./a.sh: line 4: DrShadowStack: command not found
