+++ test-files/expected-outputs/signals.expected
@@ -1,0 +1,1 @@
+./a.sh: line 4: DrShadowStack: command not found
