+++ test-files/expected-outputs/toy.expected
@@ -1,0 +1,1 @@
+./a.sh: line 4: DrShadowStack: command not found
