--- toy-examples/32-bit/exec.c
@@ -1,6 +1,0 @@
-#include <unistd.h>
-
-int main() {
-	char * args[] = {"echo", "Hello World", 0};
-	execvp(args[0], args);
-}
